server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymysql==1.1.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langchain-huggingface==0.0.1
faiss-cpu==1.7.4
pypdf==3.17.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
